// Angular/shared/enums/lookups.enums.ts
export enum LookupsEnum {
    Facebook = 6090001,
    Twitter = 6090002,
    LinkedIn = 6090003,
    YouTubeVideo = 6090004,
    OtherVideo = 6090007,
    OtherLink = 6090008,
    VideoChannel = 6090009
}

export enum LicenseStatusEnum {
    Outstanding = 1,
    Purchased = 2,
    Activated = 3,
    Cancelled = 4,
    Expired = 5,
}

export enum ProductIDEnum {
    IRPrepTool = 5077,
    IRPrepToolForEA = 5130,
    ESRCode = 5086,
    FreeGmatPrep = 5035
}

export enum AccommodationStatus {
    NewRequest = "New Request",
    DecisionMade = "Decision Made",
    UnderReview = "Under Review"
}